Add UI components and background workers for duplicate detection
- Implement reusable widgets for displaying duplicate clusters in `dup_widgets.py`.
- Create background workers for scanning and refining duplicates in `dup_workers.py`.
- Introduce file action utilities for managing duplicates in `file_actions.py`.
- Add indexing tasks for tag search functionality in `index_tasks.py`.
- Develop item delegates for rendering tag search results in `result_delegates.py`.
- Implement rendering helpers for tag display in `tag_rendering.py`.
- Create thumbnail loading tasks for tag search results in `thumbnail_tasks.py`.
- Add smoke tests for duplicate detection UI behavior in `test_dup_tab_smoke.py`.
- Implement tests for duplicate UI background workers in `test_dup_workers.py`.
